agent/pyproject.toml
[project]
name = "convoguide-agent"
version = "0.1.0"
description = "Adaptive conversational AI agent with specialist style routing"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents>=1.0.0",
    "livekit-plugins-openai>=1.0.0",
    "livekit-plugins-deepgram>=1.0.0",
    "livekit-plugins-cartesia>=1.0.0",
    "livekit-plugins-silero>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
start = "src.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
